Doc: Definimos las tareas de cada usuario

+++ backend/controllers/tareaController.js
@@ -3,7 +3,7 @@ const Tarea = require('../models/tareaModel')
 
 const getTareas = asyncHandler(async (req, res) => {
 
-    const tareas = await Tarea.find()
+    const tareas = await Tarea.find({user: req.user.id})
 
     res.status(200).json(tareas)
 })
@@ -17,7 +17,8 @@ const setTareas = asyncHandler(async(req,res) => {
     }
 
     const tarea = await Tarea.create({
-        texto: req.body.texto
+        texto: req.body.texto,
+        user: req.user.id
     })
 
     res.status(201).json(tarea)
@@ -26,10 +27,15 @@ const setTareas = asyncHandler(async(req,res) => {
 const updateTareas = asyncHandler(async(req,res) => {
 
     const tarea = await Tarea.findById(req.params.id)
-
+    //* Verificamos que la tarea exista
     if (!tarea) {
         res.status(400)
         throw new Error('Tarea no encontrada')
+    }
+    //? Verificamos q la tarea pertenezca al usuario del token 
+    if(tarea.user.toString() !== req.user.id){
+        res.status(401)
+        throw new Error('Acceso No Autorizado, la tarea no pertenece al usuario logeado')
     }
 
     const tareaModificada = await Tarea.findByIdAndUpdate(req.params.id, req.body, { new: true })
@@ -46,12 +52,16 @@ const deleteTareas = asyncHandler(async(req,res) => {
         throw new Error ('Tarea no encontrada')
     }
 
+    //? Verificamos q la tarea pertenezca al usuario del token 
+    if(tarea.user.toString() !== req.user.id){
+        res.status(401)
+        throw new Error('Acceso No Autorizado, la tarea no pertenece al usuario logeado')
+    }
+
     // await tarea.remove    //? Esta solo la borra si existe
     await tarea.deleteOne()
 
     //const tareaBorrada = await Tarea.findByIdAndDelete(req.params.id)  //? Es correcta pero ya no tiene caso q ponga cual va a borrar
-
-
     res.status(200).json({id: req.params.id})
 })
